Fix max hostname length

+++ src/Homie/Limits.hpp
@@ -8,7 +8,7 @@ namespace HomieInternals {
 
   const uint8_t MAX_WIFI_SSID_LENGTH = 32 + 1;
   const uint8_t MAX_WIFI_PASSWORD_LENGTH = 64 + 1;
-  const uint8_t MAX_HOSTNAME_LENGTH = 255 + 1;
+  const uint16_t MAX_HOSTNAME_LENGTH = 255 + 1;
 
   const uint8_t MAX_MQTT_CREDS_LENGTH = 32 + 1;
   const uint8_t MAX_MQTT_BASE_TOPIC_LENGTH = sizeof("shared-broker/username-lolipop/homie/sensors/");
